LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/replace-before-load/window-open-popup-expected.txt


FAIL No replace before load, triggered by window.open() on a non-_self window assert_equals: history.length must increase expected 2 but got 1
